[#4871] Refactoring
- Extract the same methods delegate them to RealTimeChartComponent
