+++ web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-agent-chart.component.html
@@ -1,4 +1,7 @@
-<canvas #canvas
-    (mousemove)="onMouseMove($event)"
-    (click)="onClick($event)">
-</canvas>
+<pp-new-real-time-chart
+    [activeThreadCounts]="activeThreadCounts"
+    [dataList]="dataList"
+    [timeStamp]="timeStamp"
+    [chartOption]="chartOption"
+    (outClick)="onClick($event)">
+</pp-new-real-time-chart>
