+++ web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-chart.component.css
@@ -13,13 +13,14 @@
     padding: 5px 7px;
     position: absolute;
     top: 110%;
+    width: 130px;
 }
 
 .l-tooltip-caret {
     width: 0;
     height: 0;
     position: absolute;
-    top: calc(110% - 7px);
+    top: calc(110% - 6px);
     border-bottom: 7px solid rgba(0, 0, 0, 0.8);
     border-left: 7px solid transparent;
     border-right: 7px solid transparent;
