+++ web/src/main/webapp/v2/src/app/core/components/real-time-new/new-real-time-total-chart.component.css
@@ -4,21 +4,6 @@
     width: 100%;
     height: 100%;
 }
-.l-error-template {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    left: 0;
-    top: 0;
-    justify-content: center;
-    align-items: center;
-    z-index: 2;
-}
-.l-error-text {
-    font-size: 14px;
-    font-weight: 600;
-    color: #c04e3f;
-}
 .l-legend {
     position: absolute;
     list-style: none;
@@ -68,50 +53,4 @@
 }
 .l-legend li:nth-child(5) .l-circle {
     background:#33b692
-}
-.l-tooltip {
-    background-color: rgba(0, 0, 0, 0.8);
-    color: #fff;
-    border: 0px solid rgba(0, 0, 0, 0);
-    border-radius: 5px;
-    padding: 5px 7px;
-    position: absolute;
-    top: 110%;
-    width: 130px;
-}
-
-.l-tooltip-caret {
-    width: 0;
-    height: 0;
-    position: absolute;
-    top: calc(110% - 6px);
-    border-bottom: 7px solid rgba(0, 0, 0, 0.8);
-    border-left: 7px solid transparent;
-    border-right: 7px solid transparent;
-}
-
-.l-tooltip-title {
-    font-size: 12px;
-    font-weight: 600;
-    margin-bottom: 2px;
-}
-
-.l-tooltip-body {
-    font-size: 12px;
-}
-
-.l-tooltip-body > dl {
-    display: flex;
-}
-
-.l-tooltip-body dt {
-    margin-right: 5px;
-}
-
-.l-tooltip-label {
-    display: inline-block;
-    width: 10px;
-    height: 10px;
-    margin-right: 2px;
-    box-shadow: inset 0px 0px 0px 1px rgba(255, 255, 255, 0.4);
 }
